Added the image of the schematic for better refernence
Added standard resistance values for the 500ms delay setup,
using the NPN transistor as a switch to discharge the capacitor
in ~2ms with a current of range between 0-10mA. This is well within
price and specs of the current BJTs.

Also looked into MOSFETs, but they didn't discharge the capacitor
at all(NMOS and PMOS).

NPN transistor still has to be figured out but that should be
a piece of cake hopefully..

Also, changed voltage divider values to standard, aims for
~11.7V and the resistor at the tail end of the BJT also helps
in controlling the amount of current the NPN transistor needs
to flow through it.